refactor(tests): move selenium fixtures to top level
These can be used for end to end tests as well as Find MoJ data tests.
